requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP client for API calls
aiohttp==3.9.1
httpx==0.25.2

# Image processing
Pillow==10.1.0

# Data validation and serialization
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Security
cryptography==41.0.8

# Database (optional for future use)
# sqlalchemy==2.0.23
# alembic==1.12.1

# Caching (optional)
# redis==5.0.1

# Background tasks (optional)
# celery==5.3.4